requirements.txt
crewai==0.130.0
crewai-tools==0.47.1
openai==1.13.3
langchain-openai==0.0.8
langchain-community==0.0.29
fastapi==0.110.3
uvicorn==0.30.1
pypdf==4.2.0
python-dotenv==1.0.1
